#146 Education Details

+++ admin/delete/ajax/member.php
@@ -6,7 +6,12 @@ include_once(dirname(__FILE__) . '/../../auth.php');
 
 if ($_POST['option'] == 'delete') {
     $MEMBER = new Member($_POST['id']);
-    unlink(Helper::getSitePath() . "upload/member/" . $MEMBER->profile_picture);
+
+    if ($MEMBER->profile_picture !== 'member.png') {
+        unlink(Helper::getSitePath() . "upload/member/" . $MEMBER->profile_picture);
+    }
+
+
     $result = $MEMBER->delete();
 
     if ($result) {
